refactor: Extract SessionManager tests
